preventing to charge in memory all the data to make a count

+++ fermat-p2p-plugin-communication-network-node-bitdubai/src/main/java/com/bitdubai/fermat_p2p_plugin/layer/communications/network/node/developer/bitdubai/version_1/structure/rest/services/NetworkData.java
@@ -260,7 +260,7 @@ public class NetworkData {
                 try {
                     Map<String, Object> filter = new HashMap<>();
                     filter.put("networkService.networkServiceType", networkServiceType.toString());
-                    Long count = (long) networkServiceSessionDao.count(filter);
+                    Long count = networkServiceSessionDao.count(filter);
 
                     listNetworkServicesCount.put(networkServiceType, count);
                 } catch (CantReadRecordDataBaseException e) {
